разделяет js на модули
